# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Flipeet",
      "description": "Flipeet Books E-Commerce Store Apis Documentation",
      "version": "1.0.0"
    },
    "schemes": [
      "http"
    ],
    "basePath": "/",
    "produces": [
      "application/json"
    ],
    "paths": {
      "/book/createBook": {
        "post": {
          "summary": "Create Book",
          "description": "The Books endpoint to Publish a new Books",
          "parameters": [
            {
                "name": "title",
                "in": "query",
                "description": "Title of Book.",
                "type": "String"
                
              },
              {
                "name": "user",
                "in": "query",
                "description": "id of user creating new Book.",
                "type": "String"
                
              },    {
                "name": "oldPrice",
                "in": "query",
                "description": "old Price of Book.",
                "type": "Number"
                
              },    {
                "name": "author",
                "in": "query",
                "description": "Author of Book.",
                "type": "String"
                
              },    {
                "name": "description",
                "in": "query",
                "description": "Description of Book.",
                "type": "String"
                
              },    {
                "name": "quantity",
                "in": "query",
                "description": "Quantity of Book.",
                "type": "Number"
                
              },    {
                "name": "price",
                "in": "query",
                "description": "price of Book.",
                "type": "Number"
                
              },{
                "name": "weight",
                "in": "query",
                "description": "weight of Book.",
                "type": "String"
                
              },
              {
                "name": "isbn",
                "in": "query",
                "description": "isbn of Book.",
                "type": "String"
                
              },
              {
                "name": "book_images",
                "consumes":"multipart/form-data",
                "description": "Images of Book.",
                "type": "file",
                "in": "query"
                
              },    {
                "name": "category",
                "description": "Category of Book.",
                "type": "String",
                "in": "query"
                
              },   {
                "name": "tags",
                "description": "Tags of Book.",
                "type": "Array",
                "in": "query"
                
              },    {
                "name": "status",
                "description": "Status of Book.",
                "type": "String",
                "enum": ["unpublished", "published"],
                "in": "query"
                
              },   {
                "name": "bookCondition",
                "description": "Condition of Book.",
                "type": "String",
                "in": "query"
                
              },  {
                "name": "deliveryType",
                "description": "Delivery Type of Book.",
                "type": "String",
                "enum": ["Delivery", "Meetup"],
                "in": "query"
                
              },{
                "name": "likes",
                "description": "likes of Book.",
                "type": "String",
                "in": "query"
                
              }  ,{
                "name": "publishedAt",
                "description": "likes of Book.",
                "type": "String"
                
              }    
          ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Book created Successfully"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      }, 
      "/book/editBook": {
        "post": {
          "summary": "Edit Book",
          "description": "The Books endpoint to Edit Book Details",
          "parameters": [
            {
                "name": "title",
                "description": "Title of Book.",
                "type": "String",
                "in": "query"
                
              },
              {
                "name": "user",
                "description": "id of user creating new Book.",
                "type": "String",
                "in": "query"
                
              },    {
                "name": "oldPrice",
                "description": "old Price of Book.",
                "type": "Number",
                "in": "query"
                
              },    {
                "name": "author",
                "description": "Author of Book.",
                "type": "String",
                "in": "query"
                
              },    {
                "name": "description",
                "description": "Description of Book.",
                "type": "String",
                "in": "query"
                
              },    {
                "name": "quantity",
                "description": "Quantity of Book.",
                "type": "Number",
                "in": "query"
                
              },    {
                "name": "price",
                "description": "price of Book.",
                "type": "Number",
                "in": "query"
                
              },{
                "name": "weight",
                "description": "weight of Book.",
                "type": "String",
                "in": "query"
                
              },
              {
                "name": "isbn",
                "description": "isbn of Book.",
                "type": "String",
                "in": "query"
                
              },
              {
                "name": "book_images",
                "description": "Images of Book.",
                "type": "Array",
                "in": "query"
                
              },    {
                "name": "category",
                "description": "Category of Book.",
                "type": "String",
                "in": "query"
                
              },   {
                "name": "tags",
                "description": "Tags of Book.",
                "type": "Array",
                "in": "query"
                
              },    {
                "name": "status",
                "description": "Status of Book.",
                "type": "String",
                "enum": ["unpublished", "published"],
                "in": "query"
                
              },   {
                "name": "bookCondition",
                "description": "Condition of Book.",
                "type": "String",
                "in": "query"
                
              },  {
                "name": "deliveryType",
                "description": "Delivery Type of Book.",
                "type": "String",
                "enum": ["Delivery", "Meetup"],
                "in": "query"
                
              },{
                "name": "likes",
                "description": "likes of Book.",
                "type": "String",
                "in": "query"
                
              }  ,{
                "name": "publishedAt",
                "description": "likes of Book.",
                "type": "String",
                "in": "query"
                
              }    
          ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Book Edited Successfully"
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/book/getnewbooks": {
        "get": {
          "summary": "Get New Books ",
          "description": "Get newely Posted books",
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Book get Successfully",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PriceEstimate"
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/book/getSellerNewBooks": {
        "get": {
          "summary": "Get Currently Loggedin User new Books ",
          "description": "Get user newely Posted books",
          "parameters": [
            {
                "name": "userid",
                "description": "Currently loggedin user id.",
                "type": "String",
                "in": "query"
                
              }
            ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Book get Successfully",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PriceEstimate"
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },   
      "/book/filterbooks": {
        "post": {
          "summary": "Filter Books ",
          "description": "Filter books",
          "parameters": [
            {
                "name": "title",
                "description": "title of field to filter with like author,isbn",
                "type": "String",
                "in": "query"
                
              },{
                "name": "query",
                "description": "query of field to filter with",
                "type": "String",
                "in": "query"
                
              },{
                "name": "pageNo",
                "description": "pageNo for pagination",
                "type": "String",
                "in": "query"
                
              }
            ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "data",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PriceEstimate"
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/book/getAllBooks": {
        "post": {
          "summary": "Get all Books ",
          "description": "Get all books", 
          "parameters": [
            {
                "name": "pageNo",
                "description": "Page num for pagination.",
                "type": "String",
                "in": "query"
                
              }
            ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Book get Successfully",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PriceEstimate"
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/book/getsellerbooks": {
        "get": {
          "summary": "Get Seller all Books ",
          "description": "Get Seller all Books", 
          "parameters": [
            {
                "name": "userid",
                "description": "currently online user id",
                "type": "String",
                "in": "query"
                
              }
            ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Book get Successfully",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PriceEstimate"
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/book/getSingleBook": {
        "post": {
          "summary": "Get Singlebook details ",
          "description": "Get Singlebook details", 
          "parameters": [
            {
                "name": "page",
                "description": "page number for pagination",
                "type": "String",
                "in": "query"
                
              }, {
                "name": "id",
                "description": "id of book",
                "type": "String",
                "in": "query"
                
              }
            ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Book get Successfully",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PriceEstimate"
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },"/book/searchBook": {
        "post": {
          "summary": "search Books ",
          "description": "search books", 
          "parameters": [
            {
                "name": "title",
                "description": "title is search query",
                "type": "String",
                "in": "query"
                
              },{
                "name": "option",
                "description": "option is search option to search with like isbn",
                "type": "String",
                "in": "query"
                
              }
            ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Book Found in Record",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PriceEstimate"
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/book/commentBook": {
        "post": {
          "summary": "comment Books ",
          "description": "comment books", 
          "parameters": [
            {
                "name": "userid",
                "description": "id of currently loggedin user",
                "type": "String",
                "in": "query"
                
              },{
                "name": "review",
                "description": "comment on book",
                "type": "String",
                "in": "query"
                
              },{
                "name": "bookid",
                "description": "book id on which comment created",
                "type": "String",
                "in": "query"
                
              },{
                "name": "type",
                "description": "option is search option to search with like isbn",
                "type": "String",
                "enum": ["Book Review"],
                "in": "query"
                
              }
            ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Book Found in Record",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PriceEstimate"
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/getSellerNewBooks": {
        "get": {
          "summary": "Time Estimates",
          "description": "The Time Estimates endpoint returns ETAs for all products offered at a given location, with the responses expressed as integers in seconds. We recommend that this endpoint be called every minute to provide the most accurate, up-to-date ETAs.",
          "parameters": [
            {
              "name": "start_latitude",
              "in": "query",
              "description": "Latitude component of start location.",
              "type": "number",
              "format": "double"
            },
            {
              "name": "start_longitude",
              "in": "query",
              "description": "Longitude component of start location.",
              "type": "number",
              "format": "double"
            },
            {
              "name": "customer_uuid",
              "in": "query",
              "type": "string",
              "format": "uuid",
              "description": "Unique customer identifier to be used for experience customization."
            },
            {
              "name": "product_id",
              "in": "query",
              "type": "string",
              "description": "Unique identifier representing a specific product for a given latitude & longitude."
            }
          ],
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "An array of products",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Product"
                }
              }
            },
            "401": {
              "description": "Token Authorization Denied..!",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Product"
                }
              }
            },
            "default": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
...
    }
  }